requirements.txt
Django==1.10.5
djangorestframework==3.5.4

